requirements.txt
asyncio
streamlit==1.38.0
openai==1.44.0
python-dotenv==1.0.1
pymongo==4.8.0
PyMuPDF==1.24.10
python-magic==0.4.27
olefile==0.47
pandas==2.2.2
python-docx==1.1.2
lxml==5.3.0
python-pptx==1.0.2
watchdog==5.0.2
nltk==3.9.1
xlrd==2.0.1
